+++ Assets/ProceduralStages/Main.cs
@@ -26,7 +26,7 @@ namespace ProceduralStages
         public const string PluginGUID = "Lawlzee.ProceduralStages";
         public const string PluginAuthor = "Lawlzee";
         public const string PluginName = "ProceduralStages";
-        public const string PluginVersion = "1.17.3";
+        public const string PluginVersion = "1.18";
 
         public static string SceneName = "random";
         public static string Judgement = "Judgement";